Restructured the module
