--- __init__.py
@@ -1,8 +1,0 @@
-from gym.envs.registeration import register
-
-register(
-
-id='RL_env-0', # id will be used in the main function to make an environment
-entry_point='RL_FiberTracking.envs:RLenv', # environment class
-
-)
